python/requirements.txt
asn1crypto==0.24.0
bson==0.5.8
cffi==1.12.3
cryptography==2.7
msgpack-python==0.5.6
py==1.8.0
py-bbclib~=1.5.2
pycparser==2.19
pyOpenSSL==19.0.0
pytest==3.1.3
python-dateutil==2.8.0
six==1.12.0
